# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 61bd7d167a739847d5cfd096
  title: schoolmng
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  "/student/{id}":
    delete:
      x-skaffolder-id: 61bd80657a739847d5cfd147
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 61bd7e7d7a739847d5cfd124
      x-skaffolder-resource: Student
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Student
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 61bd80647a739847d5cfd141
      x-skaffolder-name: get
      x-skaffolder-id-resource: 61bd7e7d7a739847d5cfd124
      x-skaffolder-resource: Student
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Student
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    post:
      x-skaffolder-id: 61bd80647a739847d5cfd13f
      x-skaffolder-name: update
      x-skaffolder-id-resource: 61bd7e7d7a739847d5cfd124
      x-skaffolder-resource: Student
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Student
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Student"
  "/student/findBy_courses/{key}":
    get:
      x-skaffolder-id: 61bd80927a739847d5cfd160
      x-skaffolder-name: findBy_courses
      x-skaffolder-id-resource: 61bd7e7d7a739847d5cfd124
      x-skaffolder-resource: Student
      x-skaffolder-crudAction: findBy_courses
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_courses
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_courses/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_courses
      tags:
        - Student
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
...
  "/teacher/{id}":
    delete:
      x-skaffolder-id: 61bd807e7a739847d5cfd153
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 61bd7ebd7a739847d5cfd12a
      x-skaffolder-resource: Teacher
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Teacher
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 61bd807e7a739847d5cfd14e
      x-skaffolder-name: get
      x-skaffolder-id-resource: 61bd7ebd7a739847d5cfd12a
      x-skaffolder-resource: Teacher
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Teacher
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
    post:
      x-skaffolder-id: 61bd807e7a739847d5cfd14b
      x-skaffolder-name: update
      x-skaffolder-id-resource: 61bd7ebd7a739847d5cfd12a
      x-skaffolder-resource: Teacher
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Teacher
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Teacher"
  "/teacher/findBy_courses/{key}":
    get:
      x-skaffolder-id: 61bd80927a739847d5cfd15e
      x-skaffolder-name: findBy_courses
      x-skaffolder-id-resource: 61bd7ebd7a739847d5cfd12a
      x-skaffolder-resource: Teacher
      x-skaffolder-crudAction: findBy_courses
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_courses
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_courses/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_courses
      tags:
        - Teacher
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teacher"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Student:
      x-skaffolder-id: 61bd7e7d7a739847d5cfd124
      x-skaffolder-id-db: 61bd7d167a739847d5cfd097
      x-skaffolder-id-entity: 61bd7e7d7a739847d5cfd120
      x-skaffolder-url: /student
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        DOB:
          type: number
          x-skaffolder-id-attr: 61bd7e7d7a739847d5cfd123
          x-skaffolder-type: Number
          x-skaffolder-required: true
        name:
          type: string
          x-skaffolder-id-attr: 61bd7e7d7a739847d5cfd121
          x-skaffolder-type: String
          x-skaffolder-required: true
        surname:
          type: string
          x-skaffolder-id-attr: 61bd7e7d7a739847d5cfd122
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations:
        _courses:
          x-skaffolder-id: 61bd80287a739847d5cfd13c
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 61bd7e7d7a739847d5cfd120
          x-skaffolder-ent2: 61bd7ee77a739847d5cfd12d
      required:
        - _id
        - DOB
        - name
        - surname
    Teacher:
      x-skaffolder-id: 61bd7ebd7a739847d5cfd12a
      x-skaffolder-id-db: 61bd7d167a739847d5cfd097
      x-skaffolder-id-entity: 61bd7ebd7a739847d5cfd127
      x-skaffolder-url: /teacher
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        Name:
          type: string
          x-skaffolder-id-attr: 61bd7ebd7a739847d5cfd128
          x-skaffolder-type: String
          x-skaffolder-required: true
        surname:
          type: string
          x-skaffolder-id-attr: 61bd7ebd7a739847d5cfd129
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations:
        _courses:
          x-skaffolder-id: 61bd7ff47a739847d5cfd13b
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 61bd7ebd7a739847d5cfd127
          x-skaffolder-ent2: 61bd7ee77a739847d5cfd12d
      required:
        - _id
        - Name
        - surname
...